Change model logging method to save_model
Line 66 change from artifact to direct save. writes the model directly to the file path I specify (which should be the output directory provided by Azure ML pipeline as --model_output).
This ensures Azure ML finds the model files as a "blob" at the expected location for registration in the next pipeline step.